feat(src,R): change balanced methods

+++ r-package/grf/src/src/splitting/BalancedProbabilitySplittingRule.h
@@ -48,7 +48,7 @@ namespace grf
 
   private:
     void find_best_split_value(const Data &data,
-                               size_t node, size_t var, size_t num_classes, size_t *class_counts,
+                               size_t node, size_t var, size_t num_classes, double *class_counts,
                                size_t size_node,
                                size_t min_child_size,
                                double &best_value,
@@ -68,7 +68,7 @@ namespace grf
     double imbalance_penalty;
 
     size_t *counter;
-    size_t *counter_per_class;
+    double *counter_per_class;
 
     DISALLOW_COPY_AND_ASSIGN(BalancedProbabilitySplittingRule);
   };
